Enh: Rework HTTP server part
