+++ alignak/daemons/arbiterdaemon.py
@@ -58,8 +58,7 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with Shinken.  If not, see <http://www.gnu.org/licenses/>.
 """
-This module IForArbiter and Arbiter classes used to
-communicate with HA Arbiter or some external tools
+This module provie Arbiter class used to run a arbiter daemon
 """
 import sys
 import os
@@ -73,168 +72,13 @@ import json
 from alignak.objects.config import Config
 from alignak.external_command import ExternalCommandManager
 from alignak.dispatcher import Dispatcher
-from alignak.daemon import Daemon, Interface
+from alignak.daemon import Daemon
 from alignak.log import logger
 from alignak.stats import statsmgr
 from alignak.brok import Brok
 from alignak.external_command import ExternalCommand
 from alignak.property import BoolProp
-from alignak.util import jsonify_r
-
-
-class IForArbiter(Interface):
-    """Interface for HA Arbiter. The Slave/Master arbiter can get /push conf
-
-    """
-
-    def have_conf(self, magic_hash):
-        """Does the daemon got a configuration (internal) (HTTP GET)
-
-        :param magic_hash: magic hash of configuration
-        :type magic_hash: int
-        :return: True if the arbiter has the specified conf, False otherwise
-        :rtype: bool
-        """
-        # Beware, we got an str in entry, not an int
-        magic_hash = int(magic_hash)
-        # I've got a conf and a good one
-        if self.app.cur_conf and self.app.cur_conf.magic_hash == magic_hash:
-            return True
-        else:  # I've no conf or a bad one
-            return False
-
-    def put_conf(self, conf):
-        """HTTP POST to the arbiter with the new conf (master send to slave)
-
-        :param conf: pickled the new configuration
-        :type conf:
-        :return: None
-        """
-        super(IForArbiter, self).put_conf(conf)
-        self.app.must_run = False
-    put_conf.method = 'POST'
-
-    def get_config(self):
-        """Get the managed configuration (internal) (HTTP GET)
-
-        :return: Currently managed configuration
-        :rtype: object
-        """
-        return self.app.conf
-
-    def do_not_run(self):
-        """Master tells to slave to not run (HTTP GET)
-        Master will ignore this call
-
-        :return: None
-        """
-        # If I'm the master, ignore the command
-        if self.app.is_master:
-            logger.debug("Received message to not run. "
-                         "I am the Master, ignore and continue to run.")
-        # Else, I'm just a spare, so I listen to my master
-        else:
-            logger.debug("Received message to not run. I am the spare, stopping.")
-            self.app.last_master_speack = time.time()
-            self.app.must_run = False
-    do_not_run.need_lock = False
-
-    def get_satellite_list(self, daemon_type=''):
-        """Get the satellite names sorted by type (HTTP GET)
-
-        :param daemon_type: daemon type to filter
-        :type daemon_type: str
-        :return: dict with key *daemon_type* and value list of daemon name
-        Example ::
-
-         {'poller': ['Poller1', 'Poller2']}
-
-        :rtype: dict
-        """
-        res = {}
-        for s_type in ['arbiter', 'scheduler', 'poller', 'reactionner', 'receiver',
-                       'broker']:
-            if daemon_type and daemon_type != s_type:
-                continue
-            satellite_list = []
-            res[s_type] = satellite_list
-            daemon_name_attr = s_type + "_name"
-            daemons = self.app.get_daemons(s_type)
-            for dae in daemons:
-                if hasattr(dae, daemon_name_attr):
-                    satellite_list.append(getattr(dae, daemon_name_attr))
-        return res
-
-    def what_i_managed(self):
-        """Dummy call for the arbiter
-
-        :return: {}, always
-        :rtype: dict
-        """
-        return {}
-    what_i_managed.need_lock = False
-
-    def get_all_states(self):
-        """Return all the data of satellites
-
-        :return: dict containing satellites data
-        Output looks like this ::
-
-        {'arbiter' : [{'schedproperty1':'value1' ..}, {'pollerproperty1', 'value11' ..}, ..],
-        'scheduler': [..],
-        'poller': [..],
-        'reactionner': [..],
-        'receiver': [..],
-         'broker: [..]'
-        }
-
-        :rtype: dict
-        """
-        res = {}
-        for s_type in ['arbiter', 'scheduler', 'poller', 'reactionner', 'receiver',
-                       'broker']:
-            lst = []
-            res[s_type] = lst
-            for daemon in getattr(self.app.conf, s_type + 's'):
-                cls = daemon.__class__
-                env = {}
-                all_props = [cls.properties, cls.running_properties]
-
-                for props in all_props:
-                    for prop in props:
-                        if hasattr(daemon, prop):
-                            val = getattr(daemon, prop)
-                            if prop == "realm":
-                                if hasattr(val, "realm_name"):
-                                    env[prop] = val.realm_name
-                            # give a try to a json able object
-                            try:
-                                json.dumps(val)
-                                env[prop] = val
-                            except Exception, exp:
-                                logger.debug('%s', exp)
-                lst.append(env)
-        return res
-
-    def get_objects_properties(self, table):
-        """'Dump all objects of the type in
-        [hosts, services, contacts, commands, hostgroups, servicegroups]
-
-        :param table: table name
-        :type table: str
-        :return: list all properties of all objects
-        :rtype: list
-        """
-        logger.debug('ASK:: table= %s', str(table))
-        objs = getattr(self.app.conf, table, None)
-        logger.debug("OBJS:: %s", str(objs))
-        if objs is None or len(objs) == 0:
-            return []
-        res = []
-        for obj in objs:
-            j_obj = jsonify_r(obj)
-            res.append(j_obj)
-        return res
+from alignak.http.arbiter_interface import ArbiterInterface
 
 
 class Arbiter(Daemon):
@@ -268,7 +112,7 @@ class Arbiter(Daemon):
         # Used to work out if we must still be alive or not
         self.must_run = True
 
-        self.interface = IForArbiter(self)
+        self.http_interface = ArbiterInterface(self)
         self.conf = Config()
 
     def add(self, b):
@@ -712,7 +556,6 @@ class Arbiter(Daemon):
         * Set logger
         * Log Alignak headers
         * Init daemon
-        * Register interface
         * Launch modules
         * Load retention
         * Do mainloop
@@ -740,8 +583,6 @@ class Arbiter(Daemon):
             self.look_for_early_exit()
             self.do_daemon_init_and_start()
 
-            self.uri_arb = self.http_daemon.register(self.interface)
-
             # ok we are now fully daemonized (if requested)
             # now we can start our "external" modules (if any):
             self.modules_manager.start_external_instances()
@@ -764,19 +605,20 @@ class Arbiter(Daemon):
 
         :return: None
         """
-        conf = self.new_conf
-        if not conf:
-            return
-        conf = cPickle.loads(conf)
-        self.new_conf = None
-        self.cur_conf = conf
-        self.conf = conf
-        for arb in self.conf.arbiters:
-            if (arb.address, arb.port) == (self.host, self.port):
-                self.me = arb
-                arb.is_me = lambda x: True  # we now definitively know who we are, just keep it.
-            else:
-                arb.is_me = lambda x: False  # and we know who we are not, just keep it.
+        with self.conf_lock:
+            conf = self.new_conf
+            if not conf:
+                return
+            conf = cPickle.loads(conf)
+            self.new_conf = None
+            self.cur_conf = conf
+            self.conf = conf
+            for arb in self.conf.arbiters:
+                if (arb.address, arb.port) == (self.host, self.port):
+                    self.me = arb
+                    arb.is_me = lambda x: True  # we now definitively know who we are, just keep it.
+                else:
+                    arb.is_me = lambda x: False  # and we know who we are not, just keep it.
 
     def do_loop_turn(self):
         """Loop turn for Arbiter
