+++ alignak/worker.py
@@ -366,10 +366,6 @@ class Worker:
 
         self.set_proctitle()
 
-        print "I STOP THE http_daemon", self.http_daemon
-        if self.http_daemon:
-            self.http_daemon.close_sockets()
-
         timeout = 1.0
         self.checks = []
         self.returns_queue = returns_queue
